Correciones doc hallazgos
