+++ src/components/SearchWithYearsTableReactImplementation/TableData.tsx
