+++ src/components/SearchWithYearsTableReactImplementation/types.ts
